Made asset computing async based on queues
