+++ src/async/consumers/index.ts
@@ -1,3 +1,4 @@
+export * from './asset.consumer';
 export * from './beam.consumer';
 export * from './block.consumer';
 export * from './notification.consumer';
