+++ src/utils/constants.ts
@@ -52,6 +52,7 @@ export enum ChartGroupType {
 }
 
 export enum Queues {
+    ASSETS = 'assets',
     BEAMS = 'beams',
     BLOCKS = 'blocks',
     NOTIFICATIONS = 'notifications',
@@ -59,6 +60,8 @@ export enum Queues {
 
 export enum QueueJobs {
     INGEST = 'ingest',
+    PROCESS_DAILY = 'process_daily',
+    PROCESS_WEEKLY = 'process_weekly',
     TRIGGER_VERIFY_BLOCKS_BACKWARD = 'trigger_verify_blocks_backward',
     NOTIFICATION_SOCKET = 'push_notification',
 }
